requirements.txt
colorama==0.4.5
google-cloud-storage==2.4.0
matplotlib==3.5.3
numpy==1.23.4
pandas==1.4.4
pandas-gbq==0.17.9
scikit-learn==1.3.1
statsmodels==0.13.2
google-auth==2.13.0
